Feat: The user can change the password.
From the edit page, once the user submits change password form, the code checks that the entered current password is correct.
If yes, the databse is updated with the new password and the user is redirected to the dashbaord.
If not, the edit page is rendered again with an error message.