Refactor: /user router changed to /:person.
In the login page, users and doctors can login and different actions will occur.
In this file, /user and /doctor have many common lines, they are mergerd in the same post request callback function.
depending on request.params.person, the right collection is lookedup and the user is redirected to either user's dashbaord or doctor's dashboard.